+++ src/Form/Admin/SearchPageConfigureForm.php
@@ -30,15 +30,17 @@
 namespace Search\Form\Admin;
 
 use Laminas\Form\Form;
-use Laminas\Form\Fieldset;
 use Laminas\I18n\Translator\TranslatorAwareInterface;
 use Laminas\I18n\Translator\TranslatorAwareTrait;
+use Search\Form\Element\Fields;
 
 class SearchPageConfigureForm extends Form implements TranslatorAwareInterface
 {
     use TranslatorAwareTrait;
 
     protected $formElementManager;
+
+    protected $urlViewHelper;
 
     public function init()
     {
@@ -47,6 +49,15 @@ class SearchPageConfigureForm extends Form implements TranslatorAwareInterface
         $searchPage = $this->getOption('search_page');
         $adapter = $searchPage->index()->adapter();
         $settings = $searchPage->settings();
+
+        $this->add([
+            'name' => 'save_queries',
+            'type' => 'Checkbox',
+            'options' => [
+                'label' => $translator->translate('Save queries'),
+                'info' => $translator->translate('Add ability to save your query and reload it later with result update'),
+            ],
+        ]);
 
         $this->add([
             'name' => 'facet_limit',
@@ -61,54 +72,21 @@ class SearchPageConfigureForm extends Form implements TranslatorAwareInterface
             ],
         ]);
 
-        $facets = new Fieldset('facets');
-        $facets->setLabel($translator->translate('Standard Facets'));
-        $facets->setAttribute('data-sortable', '1');
-
         $facetFields = $adapter->getAvailableFacetFields($searchPage->index());
-        $facetSelectOptions = [];
-        $weights = range(0, count($facetFields));
-        $weight_options = array_combine($weights, $weights);
-        $weight = 0;
-        foreach ($facetFields as $field) {
-            array_push($facetSelectOptions, array('value' => $field['name'], 'label' => $this->getFacetFieldLabel($field)));
-            $fieldset = new Fieldset($field['name']);
-            $fieldset->setLabel($this->getFacetFieldLabel($field));
-
-            $displayFieldset = new Fieldset('display');
-            $displayFieldset->add([
-                'name' => 'label',
-                'type' => 'Text',
-                'options' => [
-                    'label' => $translator->translate('Label'),
-                ],
-            ]);
-            $fieldset->add($displayFieldset);
-
-            $fieldset->add([
-                'name' => 'enabled',
-                'type' => 'Checkbox',
-                'options' => [
-                    'label' => $translator->translate('Enabled'),
-                ],
-            ]);
-
-            $fieldset->add([
-                'name' => 'weight',
-                'type' => 'Select',
-                'options' => [
-                    'label' => $translator->translate('Weight'),
-                    'value_options' => $weight_options,
-                ],
-                'attributes' => [
-                    'value' => $weight++,
-                ],
-            ]);
-
-            $facets->add($fieldset);
-        }
-
-        $this->add($facets);
+        $facetValueOptions = array_column($facetFields, 'label', 'name');
+        $url = $this->urlViewHelper;
+        $this->add([
+            'name' => 'facets',
+            'type' => Fields::class,
+            'options' => [
+                'label' => 'Facets', // @translate
+                'empty_option' => 'Add a facet', // @translate
+                'value_options' => $facetValueOptions,
+                'field_list_url' => $url('admin/search/facets', ['action' => 'field-list'], ['query' => ['search_page_id' => $searchPage->id()]]),
+                'field_row_url' => $url('admin/search/facets', ['action' => 'field-row'], ['query' => ['search_page_id' => $searchPage->id()]]),
+                'field_edit_sidebar_url' => $url('admin/search/facets', ['action' => 'field-edit-sidebar'], ['query' => ['search_page_id' => $searchPage->id()]]),
+            ],
+        ]);
 
         $this->add([
             'name' => 'date_range_facet_field',
@@ -119,7 +97,7 @@ class SearchPageConfigureForm extends Form implements TranslatorAwareInterface
                 'empty_option' => '',
             ],
             'attributes' => [
-              'placeholder' => 'dcterms_date_ss',
+                'placeholder' => 'dcterms_date_ss',
             ],
         ]);
 
@@ -128,47 +106,19 @@ class SearchPageConfigureForm extends Form implements TranslatorAwareInterface
         $sort_fields_fieldset->setAttribute('data-sortable', '1');
 
         $sortFields = $adapter->getAvailableSortFields($searchPage->index());
-        $weights = range(0, +count($sortFields));
-        $weight_options = array_combine($weights, $weights);
-        $weight = 0;
-        foreach ($sortFields as $field) {
-            $fieldset = new Fieldset($field['name']);
-            $fieldset->setLabel($this->getSortFieldLabel($field));
-
-            $displayFieldset = new Fieldset('display');
-            $displayFieldset->add([
-                'name' => 'label',
-                'type' => 'Text',
-                'options' => [
-                    'label' => $translator->translate('Label'),
-                ],
-            ]);
-            $fieldset->add($displayFieldset);
-
-            $fieldset->add([
-                'name' => 'enabled',
-                'type' => 'Checkbox',
-                'options' => [
-                    'label' => $translator->translate('Enabled'),
-                ],
-            ]);
-
-            $fieldset->add([
-                'name' => 'weight',
-                'type' => 'Select',
-                'options' => [
-                    'label' => $translator->translate('Weight'),
-                    'value_options' => $weight_options,
-                ],
-                'attributes' => [
-                    'value' => $weight++,
-                ],
-            ]);
-
-            $sort_fields_fieldset->add($fieldset);
-        }
-
-        $this->add($sort_fields_fieldset);
+        $sortFieldValueOptions = array_column($sortFields, 'label', 'name');
+        $this->add([
+            'name' => 'sort_fields',
+            'type' => Fields::class,
+            'options' => [
+                'label' => 'Sort fields', // @translate
+                'empty_option' => 'Add a sort field', // @translate
+                'value_options' => $sortFieldValueOptions,
+                'field_list_url' => $url('admin/search/sort-fields', ['action' => 'field-list'], ['query' => ['search_page_id' => $searchPage->id()]]),
+                'field_row_url' => $url('admin/search/sort-fields', ['action' => 'field-row'], ['query' => ['search_page_id' => $searchPage->id()]]),
+                'field_edit_sidebar_url' => $url('admin/search/sort-fields', ['action' => 'field-edit-sidebar'], ['query' => ['search_page_id' => $searchPage->id()]]),
+            ],
+        ]);
 
         $formFieldset = $this->getFormFieldset();
         if ($formFieldset) {
@@ -184,6 +134,11 @@ class SearchPageConfigureForm extends Form implements TranslatorAwareInterface
     public function getFormElementManager()
     {
         return $this->formElementManager;
+    }
+
+    public function setUrlViewHelper($urlViewHelper)
+    {
+        $this->urlViewHelper = $urlViewHelper;
     }
 
     protected function getFormFieldset()
@@ -216,7 +171,7 @@ class SearchPageConfigureForm extends Form implements TranslatorAwareInterface
         $settings = $searchPage->settings();
 
         $name = $field['name'];
-        $label = isset($field['label']) ? $field['label'] : null;
+        $label = $field['label'] ?? null;
         if (isset($settings[$settings_key][$name])) {
             $fieldSettings = $settings[$settings_key][$name];